save project2

+++ js/scriptSite.js
@@ -2,9 +2,6 @@
 
 //Function to Return the Correct Ajax Object
 
-//Ajax Load
-
-
 (function(){
     var ajaxLoadButton = document.getElementById("ajaxload-btn"),
         localLoadButton = document.getElementById("localload-btn"),
@@ -23,12 +20,11 @@
                 xhr = new ActiveXObject("Msxml2.XMLHTTP");
             }
             return xhr;
- },
-        
+        },
         ajax: function (dataUrl, outputElement, callback){
      
                 var request = site.getHttp();
-                outputElement.innerHTML = "Loading.. Please take a moment";
+                outputElement.innerHTML = "Loading.. Please take a moment.";
                 request.onreadystatechange = function() {
                     if ( request.readyState === 4 && request.status === 200 ) {
                         site.socialNetwork = JSON.parse(request.responseText);
@@ -42,10 +38,7 @@
                 }    
                 request.open("GET", dataUrl, true);
                 request.send(null);
-    },
-        
-        
-        
+            },
         search : function(event){
             var output = document.getElementById("network");    
              site.ajax('data/netSites.json', output, function (data) {
@@ -71,16 +64,19 @@
                         p_name.appendChild(p_name_text);
                         target.appendChild(p_name);
                         } // end if
-    
                     } // end for loop
-    
                 } // end count check
-            
             }); // end ajax call
         },
-        socialNetwork : function () {
+        socialNet : function () {
             var output = document.getElementById("network");
-            site.ajax('data/netSites.json', output, function (data) {
+            ajaxLoadButton.style.visibility="hidden";
+            localSaveButton.style.visibility="visible";
+            localClearButton.style.visibility="hidden";
+            site.ajax('data/netSites.json', output, site.build);
+        },
+                
+        build : function(data){ 
                 var siteList = data.socialNetworks,
                     count = siteList.length,
                     i;
@@ -100,16 +96,25 @@
                         target.appendChild(p_name);
                     } // end for loop
                 } // end count check
-            }); // end ajax call
         },   
-        //------------
          localLoad : function () {
             if (typeof(Storage) !== "undefined") 
                 {
-                    var stringObject = JSON.stringify(socialNetworks);
-                    localStorage.setItem("socialNetworks", stringObject);
+                    var storedItem = localStorage.getItem("socialNetwork");
+                    if (storedItem === null)
+                    {
+                        target.innerHTML = "No saved data found.";
+                    }
+                    else
+                    {
+                    target.innerHTML = "Data has been saved into local storage.";
+                    var convertObject = JSON.parse(storedItem);
+                    site.socialNetwork = convertObject;
+                    target.innerHTML = "Loading.. Please take a moment.";
+                    site.build(convertObject);
+                    }
                 } 
-                else 
+            else 
                 {
                     alert("Sorry, your browser does not support Web Storage...");
                 }
@@ -117,29 +122,46 @@
         localSave: function () {
             if (typeof(Storage) !== "undefined") 
                 {
-                    var storedItem = localStorage.getItem("socialNetworks");
-                    var convertObject = JSON.parse(storedItem);;
+                    if (site.socialNetwork === null)
+                    {
+                        target.innerHTML = "No data is on the page yet.";
+                    }
+                    else
+                    {
+                        target.innerHTML = "Data has been saved.";
+                        localSaveButton.style.visibility="hidden";
+                        localClearButton.style.visibility="visible";
+                    var stringObject = JSON.stringify(site.socialNetwork);
+                        
+                    localStorage.setItem("socialNetwork", stringObject);
+                    }
                 } 
-                else 
+            else 
                 {
-                    alert("Sorry, your browser does not support Web Storage...");
+                    target.innerHTML = "Sorry, your browser does not support Web Storage...";
                 }
         },
         localClear: function () {
-            localStorage.clear();
+            if (site.socialNetwork === null)
+                {
+                    target.innerHTML = "No data to be deleted in the local storage.";
+                }
+            else
+                {
+                    localStorage.clear();
+                    target.innerHTML = "Data in the local storage has been deleted.";
+                    ajaxLoadButton.style.visibility="visible";
+                    localSaveButton.style.visibility="hidden";
+                    localClearButton.style.visibility="hidden"; 
+                }
         }
-        
-        
     } // end site object
-    ajaxLoadButton.addEventListener("click", site.socialNetwork, false);
+    ajaxLoadButton.addEventListener("click", site.socialNet, false);
     searchField.addEventListener("keyup", site.search, false);
     localLoadButton.addEventListener("click", site.localLoad, false);
     localSaveButton.addEventListener("click", site.localSave, false);
     localClearButton.addEventListener("click", site.localClear, false);
 })(); // end anonymous function
-
-
-
 
 }())
